Changed large admin E2E files into smaller suites
Large E2E files limit Playwright's native sharding because tests stay grouped by file. Splitting the biggest admin files by workflow gives the shard scheduler smaller chunks without adding more CI jobs or changing test behavior.
